feat: update authentication flow and refactor user-related imports
